[BEAM-10708] Introspect beam_sql output
1. Supported implicitly executing the beam_sql query pipeline and
   introspect the output.
2. Refactored and simplified Cacheable and CacheKey usage.
3. Decoupled CacheKey calculation and some utilities from pipeline_instrument.
4. Added a simplified visualize_computed_pcoll interface.

+++ sdks/python/apache_beam/pipeline.py
@@ -96,6 +96,7 @@ from apache_beam.utils import proto_utils
 from apache_beam.utils import subprocess_server
 from apache_beam.utils.annotations import deprecated
 from apache_beam.utils.interactive_utils import alter_label_if_ipython
+from apache_beam.utils.interactive_utils import is_in_ipython
 
 if TYPE_CHECKING:
   from types import TracebackType
@@ -565,7 +566,9 @@ class Pipeline(object):
           shutil.rmtree(tmpdir)
       return self.runner.run_pipeline(self, self._options)
     finally:
-      shutil.rmtree(self.local_tempdir, ignore_errors=True)
+      if not is_in_ipython():
+        shutil.rmtree(self.local_tempdir, ignore_errors=True)
+      # else interactive beam handles the cleanup.
 
   def __enter__(self):
     # type: () -> Pipeline
